Re-unify configuration files
Unfortunately the grammar of the --config parameter is a bit of a
football right now; previously you were allowed to specify it anywhere
because there wasn't much in the way of a CLI, then when subcommands
were added it became required to specify it after the subcommand.

This change makes it so that it must be specified *before* the
subcommand.  Ideally it would just be required *somewhere* in the
command line but clap doesn't support "global" for required arguments
(and it is indeed required).

On the whole, I'm okay with the change - the net win of having a unified
config struct makes up for the ergonomics hit in my opinion, and it
unlocks some further refactoring I will do next.